chore: naming and structure
- rename `globalConfig` to `internalConfig`, for better clarity.
- move `ContextOption` to struct because it is intended only for internal use, so there is no need to export the data type.
- move `Cache` on a separated file.
